Little bit of work done
little bit more of the ground work being built still working out exactly
how I want to make the castle. considered starting with a full cube of
blocks and deleting those that don't land in the parameters for the
castle. Still working on the best way to do it by just spawning blocks
where needed.